refactored payload generation to payload generator
